+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.144 ] - 2025/01/10
+
+  * Added `SpotifyConnectDevices.UpdatePlayerDevices` method to add a list of dynamic player device entries to the `Items` collection, remove any existing dynamic devices from the collection that are not in the `playerDevices` list, and update the currently active device based on playerState.
+
 ###### [ 1.0.143 ] - 2025/01/10
 
   * Added `SpotifyConnectDevices.GetDeviceIndexByDiscoveryName` method to return the index of the `Items` collection entry that contains the specified device Zeroconf discovery results name value, or -1 if not found.